crop_chatbot_starter/requirements.txt
Flask==2.3.3
flask-cors
pandas
matplotlib
requests
python-telegram-bot==20.7
google-generativeai
gunicorn
spacy==3.7.2
